Move synchronizer contract types to partial file
Keep the public nested SkillInstallResult and SkillTargetInfo API names unchanged while moving their definitions out of the facade file. This brings ToolSkillSynchronizer under the target size after responsibility extraction.
