+++ Packages/src/Editor/Infrastructure/SkillSetup/ToolSkillSynchronizer.Types.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f5a67eed7ed46421fa649248a72cc7d1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
