+++ Packages/src/Editor/Infrastructure/SkillSetup/ToolSkillSynchronizer.cs
@@ -19,52 +19,8 @@ namespace io.github.hatayama.UnityCliLoop.Infrastructure
     /// Synchronizes skill files when tools are enabled/disabled.
     /// Removes skill directories on disable, re-installs on enable.
     /// </summary>
-    public static class ToolSkillSynchronizer
+    public static partial class ToolSkillSynchronizer
     {
-        public readonly struct SkillInstallResult
-        {
-            public readonly int AttemptedTargets;
-            public readonly int SucceededTargets;
-
-            public SkillInstallResult(int attemptedTargets, int succeededTargets)
-            {
-                AttemptedTargets = attemptedTargets;
-                SucceededTargets = succeededTargets;
-            }
-
-            public int FailedTargets => AttemptedTargets - SucceededTargets;
-            public bool IsSuccessful => FailedTargets == 0;
-        }
-
-        public readonly struct SkillTargetInfo
-        {
-            public readonly string DisplayName;
-            public readonly string DirName;
-            public readonly string InstallFlag;
-            public readonly bool HasSkillsDirectory;
-            public readonly bool HasExistingSkills;
-            public readonly bool HasDifferentLayoutSkills;
-            public readonly SkillInstallState InstallState;
-
-            public SkillTargetInfo(
-                string displayName,
-                string dirName,
-                string installFlag,
-                bool hasSkillsDirectory,
-                bool hasExistingSkills,
-                bool hasDifferentLayoutSkills = false,
-                SkillInstallState installState = SkillInstallState.Missing)
-            {
-                DisplayName = displayName;
-                DirName = dirName;
-                InstallFlag = installFlag;
-                HasSkillsDirectory = hasSkillsDirectory;
-                HasExistingSkills = hasExistingSkills;
-                HasDifferentLayoutSkills = hasDifferentLayoutSkills;
-                InstallState = installState;
-            }
-        }
-
         public static void RemoveSkillFiles(string toolName)
         {
             Debug.Assert(!string.IsNullOrEmpty(toolName), "toolName must not be null or empty");
